Started work to convert to .NET 5.0 (Core)

+++ VidCoder/Utilities/PresetConverter.cs
@@ -76,7 +76,7 @@ namespace VidCoder
 			profile.VideoOptions = hbPreset.VideoOptionExtra;
 			profile.TwoPass = hbPreset.VideoTwoPass;
 			profile.TurboFirstPass = hbPreset.VideoTurboTwoPass;
-			profile.Quality = hbPreset.VideoQualitySlider;
+			profile.Quality = (decimal)hbPreset.VideoQualitySlider;
 			profile.VideoBitrate = hbPreset.VideoAvgBitrate ?? 0;
 			profile.QsvDecode = false;
 			switch (hbPreset.VideoQualityType)
